Change color on icons in new dropdown
